fixed test errors
