+++ src/dcmri/lib.py
@@ -203,8 +203,6 @@ def fetch(dataset:str)->dict:
     return data_dict
 
 
-
-
 def influx_step(t:np.ndarray, weight:float, conc:float, dose:float, rate:float, t0:float)->np.ndarray:
     """Contrast agent flux (mmol/sec) generated by step injection.
 
@@ -812,7 +810,6 @@ def perfusion(parameter='BF', tissue='gray matter')->float:
         raise ValueError(msg)
     
     
-
 def _ellipse(array, center, axes, angle, value):
     n = array.shape[0]
     major = np.amax(axes)
@@ -1120,7 +1117,6 @@ def shepp_logan(*params, n=256, B0=3):
         return phantom
 
 
-
 def aif_parker(t, BAT:float=0.0)->np.ndarray:
     """Population AIF model as defined by `Parker et al (2006) <https://onlinelibrary.wiley.com/doi/full/10.1002/mrm.21066>`_
 
@@ -1178,7 +1174,6 @@ def aif_parker(t, BAT:float=0.0)->np.ndarray:
     pop_aif = gaussian1 + gaussian2 + sigmoid
     
     return pop_aif/1000 # convert to M
-
 
 
 def aif_tristan_rat(t,
